MAJOR UPDATE!!
The coordinates for the buildings are now on a separate js file. The WelcomeScreen background has been reworked (new colors, effects, and sizing) so it can fully be seen without stretching out. EN 1ST Floor is finished. Comfort Room options are now easier with the "NEAREST CR'" dropdown option. Animations for the graphing has been added for a smoother look.